feat: use trait for interacting with model store to fetch models
feat: new trait for interacting with model store
