+++ jams-core/src/model_store/local/mod.rs
@@ -1,0 +1,1 @@
+pub mod filesystem;
