+++ jams-serve/src/http/router.rs
@@ -29,7 +29,7 @@ mod tests {
     use crate::common::state::AppState;
     use crate::http::router::build_router;
     use jams_core::manager::ManagerBuilder;
-    use jams_core::model_store::local::LocalModelStore;
+    use jams_core::model_store::local::filesystem::LocalModelStore;
     use rayon::ThreadPoolBuilder;
     use std::sync::Arc;
