+++ jams-serve/tests/api/grpc/helper.rs
@@ -1,5 +1,5 @@
 use jams_core::manager::ManagerBuilder;
-use jams_core::model_store::local::LocalModelStore;
+use jams_core::model_store::local::filesystem::LocalModelStore;
 use jams_proto::jams_v1::model_server_client::ModelServerClient;
 use jams_proto::jams_v1::model_server_server::ModelServerServer;
 use jams_serve::common::state::AppState;
