+++ jams-serve/tests/api/http/helper.rs
@@ -1,6 +1,6 @@
 use axum::Router;
 use jams_core::manager::ManagerBuilder;
-use jams_core::model_store::local::LocalModelStore;
+use jams_core::model_store::local::filesystem::LocalModelStore;
 use jams_serve::common::state::AppState;
 use jams_serve::http::router::build_router;
 use rayon::ThreadPoolBuilder;
